upon further review, I think I don't know a thing about integrating a distribution the way LeBris has it layed out.
Also, it seems Pedersen et al. used the form dnorm(grid, observation, sd). I always thought of it the other way. If we think about it this way, we can use the minT maxT from the locfit to get an adhoc sd (max-min)/4..

Integration (hopefully) solved! I wrote a function with Google's help to integrate from minT to maxT using dnorm at each cell. Note the changes in the assignemnt in dnorm. I think we were doing it wrong... or, Pedersen et al. is wrong.